Initial move implementation to source
