+++ core/include/webview/detail_impl.hh
@@ -1,0 +1,6 @@
+#ifndef DETAIL_IMPL_HH
+#define DETAIL_IMPL_HH
+
+#include "../../src/detail/engine_base.cc"
+
+#endif
