+++ core/include/webview/webview.h
@@ -28,5 +28,6 @@
 
 #include "api.h"
 #include "c_api_impl.hh"
+#include "detail_impl.hh"
 
 #endif // WEBVIEW_H
